chore: remove outdated package
