+++ app/components/FormCreator/index.js
@@ -1,12 +1,11 @@
 import {
   Button, Chip, CircularProgress,
-  Grid, Input, MenuItem, Select,
+  Grid, Input, MenuItem, Select, NoSsr
 } from '@material-ui/core';
 import useInput from 'app/modules/hooks';
 import { customQueryService, postObjectsToClass } from 'app/services/parse';
 import { useEffect, useState } from 'react';
 import { DragDropContext } from 'react-beautiful-dnd';
-import NoSSR from 'react-no-ssr';
 import _ from 'underscore';
 import { v4 as uuid } from 'uuid';
 
@@ -115,9 +114,10 @@ function FormCreator() {
     },
     [setFormItems],
   );
+
   return (
     <div className={styles.formCreator}>
-      <NoSSR>
+      <NoSsr>
         <DragDropContext onDragEnd={onDragEnd}>
           <Grid container>
             <Grid item xs={9}>
@@ -198,7 +198,7 @@ function FormCreator() {
             </Grid>
           </Grid>
         </DragDropContext>
-      </NoSSR>
+      </NoSsr>
 
     </div>
   );
